SetupWrapper: bestVersion: Reverse before maximumBy.
Hack (?) to get around #1463 messing up local Cabal development
environments, compiling Setup.hs files with older Cabal library
versions because the system ones are used instead of the user ones.

This just reverts the order, which seems to work for me for now,
but that bug has many open questions.