WIP: Cache build directories for faster rebuilds.
We can simply build in `~/.cabal/buildcache/package-1.2.3` instead
of a temporary directory that is deleted after the build, making
use of `ghc --make`'s excellent recompilation avoidance and
flag change detection.

This is a space/time tradeoff.
The user can delete `~/.cabal/buildcache` any time if desired.

The speedup for --reinstall or installing the same package in multiple
sandboxes would be significant.
For example, a `haskell-src-exts` no-op rebuild takes 7 seconds
on my PC as compared to 132 seconds for a full build.

WIP:
Doesn't work for sandboxes yet because I don't understand how
the `maybeRenameDistDir` function works.